% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_cluster_stability.R
\name{calculate_ari}
\alias{calculate_ari}
\title{Calculate Adjusted Rand Index (ARI) between two clusterings}
\usage{
calculate_ari(labels1, labels2)
}
\arguments{
\item{labels1}{First set of cluster labels}

\item{labels2}{Second set of cluster labels}
}
\value{
Adjusted Rand Index (ARI) value
}
\description{
Calculate Adjusted Rand Index (ARI) between two clusterings
}
\keyword{internal}
