% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{check_annotation_alignment}
\alias{check_annotation_alignment}
\title{Check AnnDatR Alignment
Verifies that the expression matrix (X) rows and columns are
correctly ordered and aligned with the gene (var) and cell (obs) metadata tables.}
\usage{
check_annotation_alignment(adata)
}
\arguments{
\item{adata}{AnnDatR object.}
}
\value{
Logical TRUE if alignment is perfect, FALSE otherwise.
}
\description{
Check AnnDatR Alignment
Verifies that the expression matrix (X) rows and columns are
correctly ordered and aligned with the gene (var) and cell (obs) metadata tables.
}
\keyword{internal}
