% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_cluster_compare.R
\name{cluster_matching}
\alias{cluster_matching}
\title{Cluster matching algorithm}
\usage{
cluster_matching(cluster_comparison, total_genes)
}
\arguments{
\item{cluster_comparison}{A tibble containing pairwise cluster comparison results, including columns: cluster_A, cluster_B, n_overlap, adj_p_val.}

\item{total_genes}{Total number of genes considered in the comparison.}
}
\value{
A list containing:
\itemize{
\item matched_clusters: A tibble of matched clusters adhering to the matching rules.
\item match_score: A numeric score representing the total normalized overlap percentage.
}
}
\description{
\code{cluster_matching()} implements an algorithm to match clusters from two datasets based on significant overlaps, ensuring that there are no many-to-many matches, while still possible to have one-to-many or many-to-one matches.
}
\keyword{internal}
