% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_adata.R
\docType{data}
\name{example_adata}
\alias{example_adata}
\title{Human Protein Atlas Tissue Data (transposed and subsampled)}
\format{
An AnnDatR object with the following structure:
\describe{
\item{X}{A data frame with 78 rows and 1001 columns representing the main data matrix (e.g., gene expression values).}
\item{obs}{A data frame with 1000 rows and 11 columns containing metadata for observations (e.g., genes).}
\item{obs_names_col}{A string specifying the column used for observation names (\code{ensembl_id}).}
\item{var}{A data frame with 78 rows and 7 columns containing metadata for variables (e.g., samples).}
\item{var_names_col}{A string specifying the column used for variable names (\code{sample_id}).}
\item{uns}{A list for storing unstructured metadata (currently \code{NULL}).}
\item{obsm}{A list for storing multidimensional annotations for observations (currently \code{NULL}).}
\item{raw}{A slot for storing raw data (currently \code{NULL}).}
}
}
\source{
\url{https://www.science.org/doi/10.1126/science.1260419}
}
\usage{
example_adata
}
\description{
Human Protein Atlas tissue bulk-RNA expression data in AnnDatR format transposed and subsampled (first 1000 genes).
}
\examples{
example_adata
}
\keyword{datasets}
