% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_cluster_hulls.R
\name{get_density}
\alias{get_density}
\title{Estimate 2D density}
\usage{
get_density(x, y, h = 0.5, n = 100, lims = c(range(x), range(y)))
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{h}{bandwidth}

\item{n}{number of grid points}

\item{lims}{limits}
}
\value{
Tidy tibble with x, y (grid indices), z (density) and the
corresponding x_coord, y_coord grid coordinates.
}
\description{
Estimate 2D density
}
\keyword{internal}
