% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_annot_dbs.R
\name{hc_annot_dbs}
\alias{hc_annot_dbs}
\title{Download annotation databases for HPAclusteR}
\usage{
hc_annot_dbs(dest_dir = "databases", hpa_version = 24)
}
\arguments{
\item{dest_dir}{Destination directory to save the downloaded databases (default: "databases").}

\item{hpa_version}{Version of the Human Protein Atlas to download (default: 24).}
}
\value{
Invisible TRUE if all downloads are successful.
}
\description{
\code{hc_annot_dbs()} downloads several annotation databases commonly used in single-cell RNA-seq analysis,
including the Human Protein Atlas, PanglaoDB cell markers, Reactome pathways, and TRRUST transcriptional regulatory relationships.
The databases are downloaded to the specified destination directory and unzipped or decompressed if necessary.
}
\examples{
hc_annot_dbs()
}
