% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_distance.R
\name{hc_distance}
\alias{hc_distance}
\title{Compute distance matrix from PCA scores}
\usage{
hc_distance(AnnDatR, components = NULL, method = "spearman")
}
\arguments{
\item{AnnDatR}{An AnnDatR object containing the data with PCA results.}

\item{components}{Number of principal components to be used. If \code{NULL}, it
will be set to the first component with over 80\% explained variance.}

\item{method}{Distance metric to use (default is \code{"spearman"}). This must be
one of \code{"euclidean"}, \code{"maximum"}, \code{"manhattan"}, \code{"canberra"},
\code{"binary"}, \code{"minkowski"}, \code{"pearson"}, \code{"spearman"} or \code{"kendall"}.}
}
\value{
An AnnDatR object with the distance matrix stored in \code{uns$distance}.
}
\description{
\code{hc_distance()} computes a distance matrix based on PCA scores stored in the
AnnDatR object.
}
\examples{
# Compute distance matrix on an AnnDatR object after PCA
adata_res <- hc_pca(example_adata, components = 40)
adata_res <- hc_distance(adata_res, components = 20)
head(adata_res$uns$distance)
}
