% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_do_distance.R
\name{hc_do_distance}
\alias{hc_do_distance}
\title{Compute distance matrix from PCA scores}
\usage{
hc_do_distance(AnnDatR, components = NULL, method = "spearman")
}
\arguments{
\item{AnnDatR}{An AnnDatR object containing the data with PCA results.}

\item{components}{Number of principal components to be used. If NULL, it will be set to the first component with over 80\% explained variance.}

\item{method}{Distance metric to use (default is 'spearman'). This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman" or "kendall".}
}
\value{
Distance matrix stored within the AnnDatR object.
}
\description{
\code{hc_do_distance()} computes a distance matrix based on PCA scores stored in the AnnDatR object.
}
\examples{
# Compute distance matrix on an AnnDatR object after PCA
adata_pca <- hc_do_pca(adata_t, components = 40)
adata_dist <- hc_do_distance(adata_pca, components = 20)
head(adata_dist$uns$distance)
}
