% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_do_snn.R
\name{hc_do_snn}
\alias{hc_do_snn}
\title{Compute Shared Nearest Neighbors (SNN) Graph}
\usage{
hc_do_snn(
  AnnDatR,
  neighbors = 20,
  prune = 1/15,
  similarity = "euclidean",
  verbose = TRUE
)
}
\arguments{
\item{AnnDatR}{An AnnDatR object containing the data with distance matrix results.}

\item{neighbors}{Number of nearest neighbors to consider (default is 20).}

\item{prune}{Pruning threshold for the SNN graph (default is 1/15).}

\item{similarity}{Distance metric to use for nearest neighbor search (default is "euclidean"). Other options include "cosine", "manhattan", and "hamming".}

\item{verbose}{Logical indicating whether to print progress messages (default is TRUE).}
}
\value{
SNN graph stored within the AnnDatR object.
}
\description{
\code{hc_do_snn()} computes a Shared Nearest Neighbors (SNN) graph from the distance matrix stored in the AnnDatR object.
}
\examples{
# Compute SNN graph on an AnnDatR object after distance calculation
adata_pca <- hc_do_pca(adata_t, components = 40)
adata_dist <- hc_do_distance(adata_pca, components = 20)
adata_snn <- hc_do_snn(adata_dist, neighbors = 15)
adata_snn$uns$neighbors$snn
}
