% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_plot_umap.R
\name{hc_plot_umap}
\alias{hc_plot_umap}
\title{Plot UMAP with clusters and hulls}
\usage{
hc_plot_umap(AnnDatR, plot = c("points", "hulls", "both"))
}
\arguments{
\item{AnnDatR}{An AnnData object containing UMAP coordinates and cluster hulls.}

\item{plot}{Type of plot to create: "points" for scatter plot, "hulls" for cluster hulls, or "both" for overlaying points on hulls. Default is "points".}
}
\value{
A ggplot2 object representing the UMAP visualization.
}
\description{
\code{hc_plot_umap()} creates a UMAP plot from an AnnData object, displaying points, cluster hulls, or both.
}
\examples{
# Plot UMAP points from an AnnDatR object
adata_res <- hc_pca(example_adata, components = 40)
adata_res <- hc_distance(adata_res, components = 20)
adata_res <- hc_snn(adata_res, neighbors = 15)
adata_res <- hc_cluster_consensus(adata_res, resolution = 8, n_seeds = 20)
adata_res <- hc_umap(adata_res)
adata_res <- hc_cluster_hulls(adata_res)
hc_plot_umap(adata_res)
}
