% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_scale.R
\name{hc_scale}
\alias{hc_scale}
\title{Scale a data matrix}
\usage{
hc_scale(data, method = "sample")
}
\arguments{
\item{data}{Data to be scaled, with samples in rows and genes in columns.}

\item{method}{Scaling method: \code{"sample"} (default), \code{"gene"}, \code{"pareto"},
\code{"vector"}, \code{"min-max"}, \code{"max_sequencial"} or \code{"max"}.}
}
\value{
Scaled data matrix with genes in rows and samples in columns.
}
\description{
Scale a data matrix
}
\keyword{internal}
