% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_annot_functions.R
\name{map_ensembl_to_symbol}
\alias{map_ensembl_to_symbol}
\title{Map Ensembl IDs in a column to gene symbols and add as a new column}
\usage{
map_ensembl_to_symbol(df, geneid_col = "Gene IDs", new_col = "Gene names")
}
\arguments{
\item{df}{Data frame with a column of Ensembl IDs (slash-separated, default: "Gene IDs")}

\item{geneid_col}{Name of the column with Ensembl IDs (default: "Gene IDs")}

\item{new_col}{Name for the new gene symbol column (default: "Gene names")}
}
\value{
Data frame with an added column of gene symbols (slash-separated)
}
\description{
Map Ensembl IDs in a column to gene symbols and add as a new column
}
\keyword{internal}
