% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_annot_functions.R
\name{map_entrez_to_ensembl}
\alias{map_entrez_to_ensembl}
\title{Map Entrez IDs in a column back to Ensembl IDs}
\usage{
map_entrez_to_ensembl(df, geneid_col = "Gene IDs")
}
\arguments{
\item{df}{Data frame with a column named 'Gene IDs' (slash-separated Entrez IDs)}

\item{geneid_col}{Name of the column with Entrez IDs (default: "Gene IDs")}
}
\value{
Data frame with an added column of Ensembl IDs (slash-separated)
}
\description{
Map Entrez IDs in a column back to Ensembl IDs
}
