% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_annot_visualize.R
\name{plot_enrichment_treemap}
\alias{plot_enrichment_treemap}
\title{Plot treemap for each cluster in reduced GO terms}
\usage{
plot_enrichment_treemap(
  reduced_terms,
  ontology_pal = c(MF = "#FF6F00", CC = "#C71B00", BP = "#018EA0")
)
}
\arguments{
\item{reduced_terms}{Data frame, e.g. simp_res$reducedTerms}

\item{ontology_pal}{Named vector of ontology colors (default: c(MF="#FF6F00", CC="#C71B00", BP="#018EA0"))}
}
\value{
Named list of ggplot treemaps, one per cluster
}
\description{
Plot treemap for each cluster in reduced GO terms
}
