% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_annot_functions.R
\name{reduce_go_terms}
\alias{reduce_go_terms}
\title{Reduce GO terms}
\usage{
reduce_go_terms(go_enrichment, threshold = 0.7, verbose = TRUE)
}
\arguments{
\item{go_enrichment}{Data frame with GO enrichment results (must have
Cluster ID, Term ID, Database, Adjusted P-value)}

\item{threshold}{Similarity threshold for reduction (default: 0.7)}

\item{verbose}{Logical indicating whether to report groups that could not be
reduced (default is TRUE).}
}
\value{
List with two elements:
\itemize{
\item combined: Data frame with original and simplified GO terms
\item reducedTerms: Data frame with reduced GO terms information
}
}
\description{
Reduce GO terms
}
\keyword{internal}
