% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_annot_functions.R
\name{run_go_enrichment}
\alias{run_go_enrichment}
\title{Run GO enrichment analysis (BP, MF, CC) for all clusters using clusterProfiler::enrichGO}
\usage{
run_go_enrichment(AnnDatR, universe = NULL, verbose = TRUE)
}
\arguments{
\item{AnnDatR}{AnnDatR object containing clustering results}

\item{universe}{Character vector of background genes (default: NULL, all genes in clustering data)}

\item{verbose}{Logical indicating whether to print progress messages (default is TRUE).}
}
\value{
Tibble with enrichment results for all clusters and all GO ontologies, with Ensembl gene IDs
}
\description{
Run GO enrichment analysis (BP, MF, CC) for all clusters using clusterProfiler::enrichGO
}
\keyword{internal}
