% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_annot_functions.R
\name{run_kegg_enrichment}
\alias{run_kegg_enrichment}
\title{Run KEGG enrichment analysis (with Ensembl-to-Entrez conversion)}
\usage{
run_kegg_enrichment(AnnDatR, universe = NULL, verbose = TRUE)
}
\arguments{
\item{AnnDatR}{AnnDatR object containing clustering results}

\item{universe}{Character vector of background genes (default: NULL, all genes in clustering data)}

\item{verbose}{Logical indicating whether to print progress messages (default is TRUE).}
}
\value{
Tibble with enrichment results for all clusters
}
\description{
Run KEGG enrichment analysis (with Ensembl-to-Entrez conversion)
}
\keyword{internal}
