% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_data}
\alias{scale_data}
\title{Scale data matrix}
\usage{
scale_data(data, method = "sample")
}
\arguments{
\item{data}{Data to be scaled}

\item{method}{Scaling method: 'sample' (default), 'gene', 'pareto', 'vector', 'min-max', 'max_sequencial', or 'max'}
}
\value{
Scaled data matrix
}
\description{
Scale data matrix
}
\keyword{internal}
