% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hpa.R
\name{theme_cluster}
\alias{theme_cluster}
\title{A simple theme for visualizations}
\usage{
theme_cluster()
}
\value{
A ggplot2 theme object that can be added to plots with \code{+}.
}
\description{
A simple theme for visualizations
}
\examples{
library(ggplot2)
ggplot2::ggplot(mpg, aes(displ, hwy)) +
  ggplot2::geom_point() +
  theme_cluster()
}
