% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_hc}
\alias{theme_hc}
\title{A simple theme for visualizations}
\usage{
theme_hc()
}
\value{
A ggplot2 theme object that can be added to plots with \code{+}.
}
\description{
\code{theme_hc()} provides a minimalist ggplot2 theme with no background or grid lines. Usefull for clean, publication-ready point, bar, line and boxplots among others.
}
\examples{
library(ggplot2)
ggplot2::ggplot(mpg, aes(displ, hwy)) +
  ggplot2::geom_point() +
  theme_hc()
}
