% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_umap.R
\name{version_gte}
\alias{version_gte}
\title{Check if version v1 is greater than or equal to version v2}
\usage{
version_gte(v1, v2)
}
\arguments{
\item{v1}{version string 1}

\item{v2}{version string 2}
}
\value{
TRUE if v1 >= v2, FALSE otherwise
}
\description{
Check if version v1 is greater than or equal to version v2
}
\keyword{internal}
