% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_cluster_stability.R
\name{visualize_ari}
\alias{visualize_ari}
\title{Visualize ARI scores with histogram}
\usage{
visualize_ari(ARI_scores)
}
\arguments{
\item{ARI_scores}{A tibble containing ARI scores}
}
\value{
A list containing the histogram plot and ARI scores
}
\description{
Visualize ARI scores with histogram
}
\keyword{internal}
