% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_cluster_compare.R
\name{visualize_comparison}
\alias{visualize_comparison}
\title{Visualize cluster comparison results with a bubble heatmap}
\usage{
visualize_comparison(summary_data)
}
\arguments{
\item{summary_data}{A tibble containing cluster comparison results}
}
\value{
A buble heatmap plot
}
\description{
Visualize cluster comparison results with a bubble heatmap
}
\keyword{internal}
