% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_plot_cluster.R
\name{visualize_comparison_heatmap}
\alias{visualize_comparison_heatmap}
\title{Visualize cluster comparison results with a bubble heatmap}
\usage{
visualize_comparison_heatmap(summary_data)
}
\arguments{
\item{summary_data}{A tibble containing cluster comparison results}
}
\value{
A buble heatmap plot
}
\description{
Visualize cluster comparison results with a bubble heatmap
}
\keyword{internal}
