% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_plot_cluster.R
\name{visualize_comparison_net}
\alias{visualize_comparison_net}
\title{Visualize cluster comparison results with a network graph}
\usage{
visualize_comparison_net(matched_clusters)
}
\arguments{
\item{matched_clusters}{A tibble containing matched clusters between two datasets}
}
\value{
A network graph plot
}
\description{
Visualize cluster comparison results with a network graph
}
\keyword{internal}
