% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_genes.R
\name{cluster_genes}
\alias{cluster_genes}
\title{Cluster genes}
\usage{
cluster_genes(distance, k = 10, method, genes, id = NULL, r = 10)
}
\arguments{
\item{distance}{The distance matrix.}

\item{k}{The number of clusters.}

\item{method}{The clustering method.}

\item{genes}{The genes to cluster.}

\item{id}{An identifier for the clustering run.}

\item{r}{}
}
\value{
A list containing the clustering results, time taken, and an identifier.
}
\description{
Cluster genes
}
\examples{
# cluster_genes(distance = distance_matrix, k = 10, method = "kmeans", genes = c("gene1", "gene2"))
}
