% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_annot_databases.R
\docType{data}
\name{database_formatting_functions}
\alias{database_formatting_functions}
\title{Formatting functions for annotation databases (excluding GO)}
\format{
An object of class \code{list} of length 11.
}
\usage{
database_formatting_functions
}
\description{
Each function takes (db, db_files, cluster_genes) as arguments.
}
\keyword{internal}
