% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_consensus_cluster.R
\name{find_consensus}
\alias{find_consensus}
\title{Finds consensus clustering from multiple clusterings}
\usage{
find_consensus(all_clusterings, n, verbose = TRUE)
}
\arguments{
\item{all_clusterings}{Data frame containing clustering results from multiple runs}

\item{n}{Number of clustering runs}

\item{verbose}{Logical indicating whether to print progress messages}
}
\value{
Consensus clustering results
}
\description{
Finds consensus clustering from multiple clusterings
}
\keyword{internal}
