% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_annot_databases.R
\name{get_annot_dbs}
\alias{get_annot_dbs}
\title{Download annotation databases}
\usage{
get_annot_dbs(db_loc = "databases", hpa_version = 24)
}
\arguments{
\item{db_loc}{Destination directory to save the downloaded databases (default: "databases").}

\item{hpa_version}{Version of the Human Protein Atlas to download (default: 24).}
}
\value{
A named list mapping database IDs to file paths.
}
\description{
Download annotation databases
}
\keyword{internal}
