% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_auto_cluster.R
\name{hc_auto_cluster}
\alias{hc_auto_cluster}
\title{One-call clustering pipeline}
\usage{
hc_auto_cluster(
  AnnDatR,
  layer = NULL,
  pca_components = 50,
  pca_transform = "log1p",
  distance_components = NULL,
  distance_method = "spearman",
  snn_neighbors = 20,
  snn_similarity = "euclidean",
  cluster_resolution = 6,
  cluster_method = "louvain",
  verbose = TRUE
)
}
\arguments{
\item{AnnDatR}{An AnnDataR object.}

\item{layer}{Optional layer name within the AnnDatR object to use instead of the default data.}

\item{pca_components}{Number of principal components to compute (default is 50).}

\item{pca_transform}{Data transformation method to apply before PCA. Options are "log1p", "sqrt", or "none" (default is "log1p").}

\item{distance_components}{Number of principal components to be used. If NULL, uses Kaiser’s rule to determine the number of components.}

\item{distance_method}{Distance metric to use (default is 'spearman'). This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman" or "kendall".}

\item{snn_neighbors}{Number of nearest neighbors to consider (default is 20).}

\item{snn_similarity}{Distance metric to use for nearest neighbor search (default is "euclidean"). Other options include "cosine", "manhattan", and "hamming".}

\item{cluster_resolution}{Resolution parameter for clustering (default is 6).}

\item{cluster_method}{Clustering method to use: "louvain" (default) or "leiden".}

\item{verbose}{Logical indicating whether to print progress messages (default is TRUE).}
}
\value{
An AnnDataR object with clustering and intermediate results added. The UMAP plot of the clusters is also printed.
}
\description{
\code{hc_auto_cluster} performs a full clustering pipeline on an AnnDataR object,
including PCA, distance calculation, SNN graph construction, consensus clustering,
UMAP embedding, and cluster hull generation. It prints a UMAP plot of the
resulting clusters.
}
\examples{
adata_res <- hc_auto_cluster(example_adata)
head(adata_res$uns$consensus_clustering)
}
