% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_cluster_compare.R
\name{hc_cluster_compare}
\alias{hc_cluster_compare}
\title{Compare clusters using hypergeometric test}
\usage{
hc_cluster_compare(AnnDatR1, AnnDatR2)
}
\arguments{
\item{AnnDatR1}{First AnnDatR object}

\item{AnnDatR2}{Second AnnDatR object}
}
\value{
A tibble summarizing the pairwise cluster comparisons, including overlap, percentage overlap, p-values, and adjusted p-values.
}
\description{
\code{hc_cluster_compare()} compares clusters from two AnnDatR objects using a hypergeometric test to assess the significance of overlap between clusters.
}
\examples{
# Run clustering 1
adata_res <- hc_pca(example_adata, components = 40)
adata_res <- hc_distance(adata_res, components = 20)
adata_res <- hc_snn(adata_res, neighbors = 15)
adata_res_1 <- hc_cluster_consensus(adata_res, resolution = 6)

# Run clustering 2
adata_res_2 <- hc_cluster_consensus(adata_res, resolution = 7)

# Compare clusters between the two results
hc_cluster_compare(adata_res_1, adata_res_2)
}
