% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_cluster_hulls.R
\name{hc_cluster_hulls}
\alias{hc_cluster_hulls}
\title{Calculate UMAP cluster hulls}
\usage{
hc_cluster_hulls(
  AnnDatR,
  n = 300,
  cum_z_lim = 0.95,
  frac_lim = 0.05,
  plot_range = rep(c(min(c(AnnDatR[["obs"]][["UMAP1"]], AnnDatR[["obs"]][["UMAP2"]])),
    max(c(AnnDatR[["obs"]][["UMAP1"]], AnnDatR[["obs"]][["UMAP2"]]))), 2) * 1.05,
  poly_concavity = 1,
  poly_smoothing = 1,
  relative_bandwidth = 1/200
)
}
\arguments{
\item{AnnDatR}{An AnnDatR object containing UMAP embeddings and cluster assignments.}

\item{n}{Number of grid points for density estimation (default is 300).}

\item{cum_z_lim}{Cumulative density limit for filtering (default is 0.95).}

\item{frac_lim}{Minimum fraction limit for subcluster inclusion (default is 0.05).}

\item{plot_range}{Range of the plot in the format c(xmin, xmax, ymin, ymax) (default is 1.05 times the min and max of UMAP1 and UMAP2).}

\item{poly_concavity}{Controls the concavity of the polygons (default is 1). Higher values result in less detailed polygons.}

\item{poly_smoothing}{Controls the smoothing of the polygons (default is 1). Higher values result in less detailed polygons.}

\item{relative_bandwidth}{Relative bandwidth for density estimation as a fraction of the plot diagonal (default is 1/200).}
}
\value{
AnnDatR object with UMAP cluster hulls within the AnnDatR object.
}
\description{
\code{hc_cluster_hulls()} computes density-based cluster hulls on UMAP embeddings stored within an AnnDatR object. It identifies subclusters using DBSCAN, estimates density, filters based on cumulative density and fraction limits, and generates concave hulls for each cluster.
}
\examples{
# Calculate UMAP cluster hulls on an AnnDatR object after UMAP and clustering
adata_res <- hc_pca(example_adata, components = 40)
adata_res <- hc_distance(adata_res, components = 20)
adata_res <- hc_snn(adata_res, neighbors = 15)
adata_res <- hc_consensus_cluster(adata_res, resolution = 6.3)
adata_res <- hc_umap(adata_res)
adata_res <- hc_cluster_hulls(adata_res)
head(adata_res$uns$UMAP_hulls$hulls)
}
