% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_cluster_stability.R
\name{hc_cluster_stability}
\alias{hc_cluster_stability}
\title{Calculate ARI scores for multiple clustering seeds}
\usage{
hc_cluster_stability(AnnDatR, n_seeds = 100)
}
\arguments{
\item{AnnDatR}{AnnDatR An AnnDatR object containing clustering results.}

\item{n_seeds}{Number of different random seeds to use for clustering (default is 100).}
}
\value{
A tibble containing pairs of seeds and their corresponding ARI scores.
}
\description{
\code{hc_cluster_stability()} computes the Adjusted Rand Index (ARI) scores between clustering results obtained from multiple random seeds.
}
\examples{
# Perform clustering with multiple seeds and calculate ARI scores
adata_res <- hc_pca(example_adata, components = 40)
adata_res <- hc_distance(adata_res, components = 20)
adata_res <- hc_snn(adata_res, neighbors = 15)
adata_res <- hc_consensus_cluster(adata_res, resolution = 6.3)
hc_cluster_stability(adata_res)
}
