% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_pca.R
\name{hc_pca}
\alias{hc_pca}
\title{Perform PCA on AnnDatR object}
\usage{
hc_pca(
  AnnDatR,
  components = 50,
  transform = "log1p",
  scale_by = "sample",
  layer = NULL
)
}
\arguments{
\item{AnnDatR}{An AnnDatR object containing the data to perform PCA on.}

\item{components}{Number of principal components to compute (default is 50).}

\item{transform}{Data transformation method to apply before PCA. Options are "log1p", "sqrt", or "none" (default is "log1p").}

\item{scale_by}{Scaling method to apply to the data before PCA. Options include "sample", "gene", "pareto", "vector", "min-max", "max_sequencial", or "max" (default is "sample").}

\item{layer}{Optional layer name within the AnnDatR object to use for PCA. If NULL, the main data matrix is used (default is NULL).}
}
\value{
PCA object and scores are stored within the AnnDatR object.
}
\description{
\code{hc_pca()} performs Principal Component Analysis (PCA) on the data contained within an AnnDatR object.
}
\examples{
# Run PCA on an AnnDatR object
adata_res <- hc_pca(adata_t, components = 40)

# PCA results
adata_res$obsm$X_pca

# Explained variance
adata_res$uns$pca
}
