% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_scale.R
\name{hc_scale}
\alias{hc_scale}
\title{Scale data matrix}
\usage{
hc_scale(data, method = "sample")
}
\arguments{
\item{data}{Data to be scaled}

\item{method}{Scaling method: 'sample' (default), 'gene', 'pareto', 'vector', 'min-max', 'max_sequencial', or 'max'}
}
\value{
Scaled data matrix
}
\description{
Scale data matrix
}
\keyword{internal}
